requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2
python-dotenv>=0.19.0
pydantic-settings>=2.0.0
PyMuPDF>=1.18.0
Pillow>=8.0.0 